Add QA helperFunction file to consolidate/test functions
