--- client/src/helperFunctions.js
@@ -1,5 +1,0 @@
-module.exports.QAhelpers = {
-
-  makeFilterable =
-
-}
